Get track data from icecast instead of liquidsoap
Some clients send metadata in iso-8859-1 instead of utf8. liquidsoap is
supposed to detect and convert that (according to documentation) but it
does not work reliably. So let liquidsoap (for now) instead trigger a
fetch of the trackdata from icecast. Upcoming icecast version is supposed
to have a metadata hook so then we can do that cleaner.